pyproject.toml
[project]
name = "gitlab-docs-mcp"
version = "0.1.0"
description = "A Model Context Protocol server for GitLab documentation"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Nuno Lima", email = "nunojorgelima@gmail.com"}
]
keywords = ["mcp", "gitlab", "documentation", "model-context-protocol"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "mcp",
    "python-frontmatter",
    "markdown-it-py",
    "sqlite-utils",
]

[project.urls]
Homepage = "https://github.com/nunolima/gitlab-docs-mcp"
Repository = "https://github.com/nunolima/gitlab-docs-mcp"
Issues = "https://github.com/nunolima/gitlab-docs-mcp/issues"

[project.scripts]
gitlab-docs-mcp = "server.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server", "indexer"]
